1. Сделал нормальный logic_menu_inline.py, т.к до этого был файл logic_menu_second.py в котором была inline кнопки 2. Почему - то в файле  logic_menu_inline.py  была создание ОБЫЧНОЙ кнопки, что очень странно и непонятно. Сделал нормальную кнопку в inline (взял пример из math_training), а обычную кнопку создал в logic_menu 3. Начал делать комментарии, но нужно чётко всё описывать
